requirements/app.txt
fastapi==0.93.0
starlette==0.17.1
requests==2.27.1
mongoengine==0.24.1
PyJWT==2.3.0
uvicorn==0.17.6
pymongo==4.1.1
pendulum==2.1.2
pydantic==1.10.6
loguru==0.6.0